Separate out user menu to their own folder and components
