+++ packages/react/src/components/header/header.tsx
@@ -13,21 +13,8 @@ import { Link } from "react-router-dom";
 import { SearchBar } from "./searchbar/components/SearchBar";
 import clsx from "clsx";
 import { Logo } from "./Logo";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from "@/shadcn/ui/dropdown-menu";
-import { useAuth } from "@/hooks/useAuth";
-import { userAtom } from "@/store/auth";
 import { useState } from "react";
-import { UserMenuItem } from "./userMenu/components/UserMenuItem";
-import {
-  CalendarIcon,
-  ExitIcon,
-  GearIcon,
-  PersonIcon,
-} from "@radix-ui/react-icons";
+import { UserMenu } from "./userMenu/components/UserMenu";
 interface HeaderProps
   extends React.DetailedHTMLProps<
     React.HTMLAttributes<HTMLElement>,
@@ -105,67 +92,4 @@ export function Header({ id }: HeaderProps) {
       {!isSearching && <UserMenu />}
     </header>
   );
-}
-export function UserMenu() {
-  const { t } = useTranslation();
-  const user = useAtomValue(userAtom);
-  const { logout } = useAuth();
-
-  const mockUserMenuItems = [
-    {
-      key: "accountSettings",
-      value: "Account Settings",
-      fn: () => mockNavigate("accountSettings"),
-      icon: <PersonIcon />,
-    },
-    {
-      key: "settings",
-      value: "Settings",
-      fn: () => mockNavigate("settings"),
-      icon: <GearIcon />,
-    },
-    {
-      key: "iCalFeed",
-      value: "ICal Feed",
-      fn: () => mockNavigate("iCalFeed"),
-      icon: <CalendarIcon />,
-    },
-    { key: "logout", value: "Logout", fn: logout, icon: <ExitIcon /> },
-  ];
-
-  if (!user) {
-    return (
-      <Button asChild>
-        <Link to="/login">{t("component.mainNav.login")}</Link>
-      </Button>
-    );
-  }
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger className="mx-2 w-8 shrink-0 overflow-hidden rounded-full">
-        <img
-          src={`https://api.dicebear.com/7.x/shapes/svg?seed=${user.id}`}
-          alt="User avatar"
-        />
-      </DropdownMenuTrigger>
-      <DropdownMenuContent>
-        {/* user profile */}
-        {mockUserMenuItems.map((item) => {
-          return (
-            <UserMenuItem
-              key={item.key + item.value}
-              item={{ key: item.key, value: item.value, icon: item.icon }}
-              onClick={item.fn}
-              // add specific icon
-            />
-          );
-        })}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function mockNavigate(destination: string) {
-  console.log(destination);
 }
