+++ packages/react/src/components/header/userMenu/components/UserMenu.tsx
@@ -1,0 +1,107 @@
+import { useAuth } from "@/hooks/useAuth";
+import { userAtom } from "@/store/auth";
+import {
+  DropdownMenu,
+  DropdownMenuTrigger,
+  DropdownMenuContent,
+  DropdownMenuSeparator,
+} from "@radix-ui/react-dropdown-menu";
+import {
+  PersonIcon,
+  GearIcon,
+  CalendarIcon,
+  ExitIcon,
+} from "@radix-ui/react-icons";
+import { useAtomValue } from "jotai";
+import { useTranslation } from "react-i18next";
+import { Link } from "react-router-dom";
+import { UserMenuItem } from "./UserMenuItem";
+import { Button } from "@/shadcn/ui/button";
+import { DropdownMenuLabel } from "@/shadcn/ui/dropdown-menu";
+
+export function UserMenu() {
+  const { t } = useTranslation();
+  const user = useAtomValue(userAtom);
+  const { logout } = useAuth();
+
+  const mockUserMenuItems = [
+    {
+      key: "accountSettings",
+      value: "Account Settings",
+      fn: () => mockNavigate("accountSettings"),
+      icon: <PersonIcon />,
+    },
+    {
+      key: "settings",
+      value: "Settings",
+      fn: () => mockNavigate("settings"),
+      icon: <GearIcon />,
+    },
+    {
+      key: "iCalFeed",
+      value: "ICal Feed",
+      fn: () => mockNavigate("iCalFeed"),
+      icon: <CalendarIcon />,
+    },
+    { key: "logout", value: "Logout", fn: logout, icon: <ExitIcon /> },
+  ];
+
+  if (!user) {
+    return (
+      <Button asChild>
+        <Link to="/login">{t("component.mainNav.login")}</Link>
+      </Button>
+    );
+  }
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger className="z-30 mx-2 w-8 shrink-0 overflow-hidden rounded-full bg-base-2">
+        <img
+          src={`https://api.dicebear.com/7.x/shapes/svg?seed=${user.id}`}
+          alt="User avatar"
+        />
+      </DropdownMenuTrigger>
+      <DropdownMenuContent className="z-30 bg-base-2">
+        {/* user profile */}
+        <div>
+          <div className="p-4">
+            <DropdownMenuLabel>{user.username}</DropdownMenuLabel>
+            <div className=" flex flex-row gap-2 px-2 py-1.5 ">
+              <div
+                className={user.google_id ? "text-secondary-11" : "text-base"}
+              >
+                <div className="i-mdi:google p-1 text-xs" />
+              </div>
+              <div
+                className={user.discord_id ? "text-secondary-11" : "text-base"}
+              >
+                <div className="i-mdi:discord p-1 text-xs" />
+              </div>
+              <div
+                className={user.twitter_id ? "text-secondary-11" : "text-base"}
+              >
+                <div className="i-mdi:twitter p-1 text-xs" />
+              </div>
+            </div>
+          </div>
+        </div>
+        <DropdownMenuSeparator />
+        {mockUserMenuItems.map((item) => {
+          return (
+            <UserMenuItem
+              key={item.key + item.value}
+              item={{ key: item.key, value: item.value, icon: item.icon }}
+              onClick={item.fn}
+              // add specific icon
+            />
+          );
+        })}
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
+
+function mockNavigate(destination: string) {
+  console.log(destination);
+}
